+++ RELEASES.md
@@ -4,6 +4,8 @@
 - Bump golang version to v1.23.6
 - Bump golangci-lint to v1.63 and add linters
 
+- Moved atomic codebase from plugin/evm to plugin/evm/atomic package
+
 ## [v0.14.1](https://github.com/ava-labs/coreth/releases/tag/v0.14.1)
 
 - Removed deprecated `ExportKey`, `ExportAVAX`, `Export`, `ImportKey`, `ImportAVAX`, `Import` APIs